requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
env==0.1.0
kombu==5.3.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
vine==5.0.0
wcwidth==0.2.6
django-celery-beat==2.5.0
redis==4.5.4